data-ml-pipelines/projects/vendor_ranking/hf_transformers/requirements.txt
transformers==4.33.3
tokenizers==0.13.3
torch==2.3.1
accelerate==0.32.1
google-cloud-bigquery-storage
google-cloud-bigquery
google-cloud-storage
pandas
datasets==2.20.0
structlog==24.1.0
python-dotenv==1.0.1
mlflow==2.16.0
mlserver==1.6.0
mlserver-mlflow==1.6.0
newrelic==9.13.0
pydantic==2.7.1
pydantic_settings==2.2.1
feast==0.40.0
db-dtypes
polars==1.10.0
